+++ src/Bundle/theme/index.php
@@ -1,8 +1,10 @@
 <?php
 
+use Rocket\Helper\Theme;
+
 if ( class_exists('Timber') and !WP_DIRECT_LOADING )
 {
-    $theme = \Rocket\Model\Theme::getInstance();
+    $theme = Theme::getInstance();
     $theme->run();
 }
 else